requirements.txt
asgiref==3.2.3
asyncio==3.4.3
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
lxml==4.4.2
pytz==2019.3
spyne==2.13.2a0
sqlparse==0.3.0
suds-jurko==0.6
